Created two new AEC Microintegrartions
AEC Automatic Surchage does the exact opposite of the automatic
discount from valanx and lets you apply an extra charge automatically (%
or fixed value).

Surcharge Multiply lets you do the same but to pick a number of times to
do that.  So you could set a £10 surcharge but apply 3 of these (will be
based on a user specific multiplier using rewrite) to add £30.